+++ src/lib/ast/transform/scale.ts
@@ -4,7 +4,7 @@ import {identity, multiply} from "./utils.ts";
 export function scaleX(x: number, from: Matrix): Matrix {
 
     const matrix = identity();
-    matrix[0 * 4 + 0] = x;
+    matrix[0] = x;
 
     return multiply(from, matrix);
 }
@@ -13,7 +13,7 @@ export function scaleY(y: number, from: Matrix): Matrix {
 
     const matrix = identity();
 
-    matrix[1 * 4 + 1] = y;
+    matrix[4 + 1] = y;
 
     return multiply(from, matrix);
 }
@@ -30,8 +30,8 @@ export function scaleZ(z: number, from: Matrix): Matrix {
 export function scale(x: number, y: number, from: Matrix): Matrix {
 
     const matrix = identity();
-    matrix[0 * 4 + 0] = x;
-    matrix[1 * 4 + 1] = y;
+    matrix[0] = x;
+    matrix[4 + 1] = y;
 
     return multiply(from, matrix);
 }
@@ -39,8 +39,8 @@ export function scale(x: number, y: number, from: Matrix): Matrix {
 export function scale3d(x: number, y: number, z: number, from: Matrix): Matrix {
 
     const matrix = identity();
-    matrix[0 * 4 + 0] = x;
-    matrix[1 * 4 + 1] = y;
+    matrix[0] = x;
+    matrix[4 + 1] = y;
     matrix[2 * 4 + 2] = z;
 
     return multiply(from, matrix);