+++ src/lib/ast/transform/skew.ts
@@ -5,7 +5,7 @@ export function skewX(x: number, from: Matrix): Matrix {
 
     const matrix: Matrix = identity();
 
-    matrix[1* 4 +0] = Math.tan(x);
+    matrix[4] = Math.tan(x);
 
     return multiply(from, matrix);
 }
@@ -14,7 +14,7 @@ export function skewY(y: number, from: Matrix): Matrix {
 
     const matrix: Matrix = identity();
 
-    matrix[0* 4 +1] = Math.tan(y);
+    matrix[1] = Math.tan(y);
     return multiply(from, matrix);
 }
 
@@ -23,11 +23,11 @@ export function skew(values: [number] | [number, number], from: Matrix): Matrix 
 
     const matrix: Matrix = identity();
 
-    matrix[1* 4 +0] = Math.tan(values[0]);
+    matrix[4] = Math.tan(values[0]);
 
     if (values.length > 1) {
 
-        matrix[0* 4 +1] = Math.tan(values[1]!);
+        matrix[1] = Math.tan(values[1]!);
     }
 
     return multiply(from, matrix);