+++ src/lib/ast/transform/translate.ts
@@ -5,7 +5,7 @@ export function translateX(x: number, from: Matrix): Matrix {
 
     const matrix: Matrix = identity();
 
-    matrix[3 * 4 +0] = x;
+    matrix[3 * 4] = x;
 
     return multiply(from, matrix) as Matrix;
 }
@@ -13,7 +13,7 @@ export function translateX(x: number, from: Matrix): Matrix {
 export function translateY(y: number, from: Matrix): Matrix {
 
     const matrix: Matrix = identity();
-    matrix[3* 4 +1] = y;
+    matrix[3 * 4 + 1] = y;
 
     return multiply(from, matrix) as Matrix;
 }
@@ -21,7 +21,7 @@ export function translateY(y: number, from: Matrix): Matrix {
 export function translateZ(z: number, from: Matrix): Matrix {
 
     const matrix: Matrix = identity();
-    matrix[3* 4 +2] = z;
+    matrix[3 * 4 + 2] = z;
 
     return multiply(from, matrix) as Matrix;
 }
@@ -29,8 +29,8 @@ export function translateZ(z: number, from: Matrix): Matrix {
 export function translate(translate: [number] | [number, number], from: Matrix): Matrix {
 
     const matrix: Matrix = identity();
-    matrix[3* 4 +0] = translate[0];
-    matrix[3* 4 +1] = translate[1] ?? 0;
+    matrix[3 * 4] = translate[0];
+    matrix[3 * 4 + 1] = translate[1] ?? 0;
 
     return multiply(from, matrix) as Matrix;
 }
@@ -38,9 +38,9 @@ export function translate(translate: [number] | [number, number], from: Matrix):
 export function translate3d(translate: [number, number, number], from: Matrix): Matrix {
 
     const matrix: Matrix = identity();
-    matrix[3* 4 +0] = translate[0];
-    matrix[3* 4 +1] = translate[1];
-    matrix[3* 4 +2] = translate[2];
+    matrix[3 * 4] = translate[0];
+    matrix[3 * 4 + 1] = translate[1];
+    matrix[3 * 4 + 2] = translate[2];
 
     return multiply(from, matrix) as Matrix;
 }