+++ src/web.ts
@@ -45,7 +45,7 @@ export {
     ValidationLevel,
     ColorType,
     SourceMap,
-    WalkerValueEvent,
+    WalkerEvent,
     WalkerOptionEnum,
 } from './lib/index.ts';
 export {FeatureWalkMode} from './lib/ast/features/type.ts';