+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - [x] make streaming optional #109
 - [x] patch at-rule syntax for @font-feature-value #110
+- [x] support percentage in transform() and scale() #111
+- [x] allow arrays in visitor definition #112
 
 ## v1.3.3